+++ src/cmd/browse/shared/mod.rs
@@ -14,3 +14,6 @@
 
 mod responses;
 pub use responses::*;
+
+mod state;
+pub use state::*;
