+++ src/cmd/browse/views/card_cloze.rs
@@ -28,10 +28,10 @@ use crate::cmd::browse::render::render_config;
 use crate::cmd::browse::render::render_family_revealed;
 use crate::cmd::browse::render::render_history;
 use crate::cmd::browse::render::source_rows;
+use crate::cmd::browse::shared::BrowseState;
 use crate::cmd::browse::shared::error_response;
 use crate::cmd::browse::shared::internal_error_response;
 use crate::cmd::browse::shared::ok_response;
-use crate::cmd::browse::state::BrowseState;
 use crate::error::ErrorReport;
 use crate::error::Fallible;
 use crate::error::fail;
