+++ src/cmd/browse/views/index.rs
@@ -18,9 +18,9 @@ use axum::response::Html;
 
 use crate::cmd::browse::layout::Selection;
 use crate::cmd::browse::layout::columns_page;
+use crate::cmd::browse::shared::BrowseState;
 use crate::cmd::browse::shared::internal_error_response;
 use crate::cmd::browse::shared::ok_response;
-use crate::cmd::browse::state::BrowseState;
 
 pub async fn index_handler(State(state): State<BrowseState>) -> (StatusCode, Html<String>) {
     let selection = Selection {
